+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to
 - ✨(frontend) add configurable documentation menu item
 - ✨(frontend) allow promoting authenticated participants
 - ✨(frontend) introduce an "unauthenticated" participant badge
+- ✨(backend) add roomkit viewset to start a room without WebRTC join
 
 ### Changed